requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0

# Data processing
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0

# Utilities
requests>=2.31.0

# Optional: Jupyter for analysis
# jupyter>=1.0.0
# matplotlib>=3.7.0
# seaborn>=0.12.0
